Reorganize Volume from Ensemble
