+++ relentless/core/__init__.py
@@ -4,3 +4,5 @@ from . import math
 from .math import *
 from . import variable
 from .variable import *
+from . import volume
+from .volume import *
